feat(BUX-291): split model_transactions into smaller files (db_model_transaction, tx_repository, tx_service)
