+++ model_transactions.go
@@ -3,14 +3,10 @@ package bux
 import (
 	"context"
 	"encoding/hex"
-	"errors"
 
-	"github.com/BuxOrg/bux/chainstate"
-	"github.com/BuxOrg/bux/notifications"
 	"github.com/BuxOrg/bux/taskmanager"
 	"github.com/BuxOrg/bux/utils"
 	"github.com/libsv/go-bt/v2"
-	"github.com/mrz1836/go-datastore"
 )
 
 // TransactionBase is the same fields share between multiple transaction models
@@ -135,202 +131,11 @@ func newTransactionFromIncomingTransaction(incomingTx *IncomingTransaction) *Tra
 	return tx
 }
 
-// getTransactionByID will get the model from a given transaction ID
-func getTransactionByID(ctx context.Context, xPubID, txID string, opts ...ModelOps) (*Transaction, error) {
-	// Construct an empty tx
-	tx := newTransaction("", opts...)
-	tx.ID = txID
-	tx.XPubID = xPubID
-
-	// Get the record
-	if err := Get(ctx, tx, nil, false, defaultDatabaseReadTimeout, false); err != nil {
-		if errors.Is(err, datastore.ErrNoResults) {
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	return tx, nil
-}
-
 // setXPubID will set the xPub ID on the model
 func (m *Transaction) setXPubID() {
 	if len(m.rawXpubKey) > 0 && len(m.XPubID) == 0 {
 		m.XPubID = utils.Hash(m.rawXpubKey)
 	}
-}
-
-// getTransactions will get all the transactions with the given conditions
-func getTransactions(ctx context.Context, metadata *Metadata, conditions *map[string]interface{},
-	queryParams *datastore.QueryParams, opts ...ModelOps,
-) ([]*Transaction, error) {
-	modelItems := make([]*Transaction, 0)
-	if err := getModelsByConditions(ctx, ModelTransaction, &modelItems, metadata, conditions, queryParams, opts...); err != nil {
-		return nil, err
-	}
-
-	return modelItems, nil
-}
-
-// getTransactionsAggregate will get a count of all transactions per aggregate column with the given conditions
-func getTransactionsAggregate(ctx context.Context, metadata *Metadata, conditions *map[string]interface{},
-	aggregateColumn string, opts ...ModelOps,
-) (map[string]interface{}, error) {
-	modelItems := make([]*Transaction, 0)
-	results, err := getModelsAggregateByConditions(
-		ctx, ModelTransaction, &modelItems, metadata, conditions, aggregateColumn, opts...,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return results, nil
-}
-
-// getTransactionsCount will get a count of all the transactions with the given conditions
-func getTransactionsCount(ctx context.Context, metadata *Metadata, conditions *map[string]interface{},
-	opts ...ModelOps,
-) (int64, error) {
-	return getModelCountByConditions(ctx, ModelTransaction, Transaction{}, metadata, conditions, opts...)
-}
-
-// getTransactionsCountByXpubID will get the count of all the models for a given xpub ID
-func getTransactionsCountByXpubID(ctx context.Context, xPubID string, metadata *Metadata,
-	conditions *map[string]interface{}, opts ...ModelOps,
-) (int64, error) {
-	dbConditions := processDBConditions(xPubID, conditions, metadata)
-
-	return getTransactionsCountInternal(ctx, dbConditions, opts...)
-}
-
-// getTransactionsByXpubID will get all the models for a given xpub ID
-func getTransactionsByXpubID(ctx context.Context, xPubID string,
-	metadata *Metadata, conditions *map[string]interface{},
-	queryParams *datastore.QueryParams, opts ...ModelOps,
-) ([]*Transaction, error) {
-	dbConditions := processDBConditions(xPubID, conditions, metadata)
-
-	return getTransactionsInternal(ctx, dbConditions, xPubID, queryParams, opts...)
-}
-
-func processDBConditions(xPubID string, conditions *map[string]interface{},
-	metadata *Metadata,
-) map[string]interface{} {
-	dbConditions := map[string]interface{}{
-		"$or": []map[string]interface{}{{
-			"xpub_in_ids": xPubID,
-		}, {
-			"xpub_out_ids": xPubID,
-		}},
-	}
-
-	// check for direction query
-	if conditions != nil && (*conditions)["direction"] != nil {
-		direction := (*conditions)["direction"].(string)
-		if direction == string(TransactionDirectionIn) {
-			dbConditions["xpub_output_value"] = map[string]interface{}{
-				xPubID: map[string]interface{}{
-					"$gt": 0,
-				},
-			}
-		} else if direction == string(TransactionDirectionOut) {
-			dbConditions["xpub_output_value"] = map[string]interface{}{
-				xPubID: map[string]interface{}{
-					"$lt": 0,
-				},
-			}
-		} else if direction == string(TransactionDirectionReconcile) {
-			dbConditions["xpub_output_value"] = map[string]interface{}{
-				xPubID: 0,
-			}
-		}
-		delete(*conditions, "direction")
-	}
-
-	if metadata != nil && len(*metadata) > 0 {
-		and := make([]map[string]interface{}, 0)
-		if _, ok := dbConditions["$and"]; ok {
-			and = dbConditions["$and"].([]map[string]interface{})
-		}
-		for key, value := range *metadata {
-			condition := map[string]interface{}{
-				"$or": []map[string]interface{}{{
-					metadataField: Metadata{
-						key: value,
-					},
-				}, {
-					xPubMetadataField: map[string]interface{}{
-						xPubID: Metadata{
-							key: value,
-						},
-					},
-				}},
-			}
-			and = append(and, condition)
-		}
-		dbConditions["$and"] = and
-	}
-
-	if conditions != nil && len(*conditions) > 0 {
-		and := make([]map[string]interface{}, 0)
-		if _, ok := dbConditions["$and"]; ok {
-			and = dbConditions["$and"].([]map[string]interface{})
-		}
-		and = append(and, *conditions)
-		dbConditions["$and"] = and
-	}
-
-	return dbConditions
-}
-
-// getTransactionsInternal get all transactions for the given conditions
-// NOTE: this function should only be used internally, it allows to query the whole transaction table
-func getTransactionsInternal(ctx context.Context, conditions map[string]interface{}, xPubID string,
-	queryParams *datastore.QueryParams, opts ...ModelOps,
-) ([]*Transaction, error) {
-	var models []Transaction
-	if err := getModels(
-		ctx,
-		NewBaseModel(ModelTransaction, opts...).Client().Datastore(),
-		&models,
-		conditions,
-		queryParams,
-		defaultDatabaseReadTimeout,
-	); err != nil {
-		if errors.Is(err, datastore.ErrNoResults) {
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	// Loop and enrich
-	transactions := make([]*Transaction, 0)
-	for index := range models {
-		models[index].enrich(ModelTransaction, opts...)
-		models[index].XPubID = xPubID
-		tx := &models[index]
-		transactions = append(transactions, tx)
-	}
-
-	return transactions, nil
-}
-
-// getTransactionsCountInternal get a count of all transactions for the given conditions
-func getTransactionsCountInternal(ctx context.Context, conditions map[string]interface{},
-	opts ...ModelOps,
-) (int64, error) {
-	count, err := getModelCount(
-		ctx,
-		NewBaseModel(ModelNameEmpty, opts...).Client().Datastore(),
-		Transaction{},
-		conditions,
-		defaultDatabaseReadTimeout,
-	)
-	if err != nil {
-		return 0, err
-	}
-
-	return count, nil
 }
 
 // UpdateTransactionMetadata will update the transaction metadata by xPubID
@@ -361,35 +166,6 @@ func (m *Transaction) UpdateTransactionMetadata(xPubID string, metadata Metadata
 // GetModelName will get the name of the current model
 func (m *Transaction) GetModelName() string {
 	return ModelTransaction.String()
-}
-
-// GetModelTableName will get the db table name of the current model
-func (m *Transaction) GetModelTableName() string {
-	return tableTransactions
-}
-
-// Save will save the model into the Datastore
-func (m *Transaction) Save(ctx context.Context) (err error) {
-	// Prepare the metadata
-	if len(m.Metadata) > 0 {
-		// set the metadata to be xpub specific, but only if we have a valid xpub ID
-		if m.XPubID != "" {
-			// was metadata set via opts ?
-			if m.XpubMetadata == nil {
-				m.XpubMetadata = make(XpubMetadata)
-			}
-			if _, ok := m.XpubMetadata[m.XPubID]; !ok {
-				m.XpubMetadata[m.XPubID] = make(Metadata)
-			}
-			for key, value := range m.Metadata {
-				m.XpubMetadata[m.XPubID][key] = value
-			}
-		} else {
-			m.DebugLog("xPub id is missing from transaction, cannot store metadata")
-		}
-	}
-
-	return Save(ctx, m)
 }
 
 // GetID will get the ID
@@ -443,251 +219,6 @@ func (m *Transaction) getValues() (outputValue uint64, fee uint64) {
 	// remove the fee from the value
 	if outputValue > fee {
 		outputValue -= fee
-	}
-
-	return
-}
-
-// BeforeCreating will fire before the model is being inserted into the Datastore
-func (m *Transaction) BeforeCreating(ctx context.Context) error {
-	if m.beforeCreateCalled {
-		m.DebugLog("skipping: " + m.Name() + " BeforeCreating hook, because already called")
-		return nil
-	}
-
-	m.DebugLog("starting: " + m.Name() + " BeforeCreating hook...")
-
-	// Test for required field(s)
-	if len(m.Hex) == 0 {
-		return ErrMissingFieldHex
-	}
-
-	// Set the xPubID
-	m.setXPubID()
-
-	// Set the ID - will also parse and verify the tx
-	err := m.setID()
-	if err != nil {
-		return err
-	}
-
-	// 	m.xPubID is the xpub of the user registering the transaction
-	if len(m.XPubID) > 0 && len(m.DraftID) > 0 {
-		// Only get the draft if we haven't already
-		if m.draftTransaction == nil {
-			if m.draftTransaction, err = getDraftTransactionID(
-				ctx, m.XPubID, m.DraftID, m.GetOptions(false)...,
-			); err != nil {
-				return err
-			} else if m.draftTransaction == nil {
-				return ErrDraftNotFound
-			}
-		}
-	}
-
-	// Validations and broadcast config check
-	if m.draftTransaction != nil {
-
-		// No config set? Use the default from the client
-		if m.draftTransaction.Configuration.Sync == nil {
-			m.draftTransaction.Configuration.Sync = m.Client().DefaultSyncConfig()
-		}
-
-		// Create the sync transaction model
-		sync := newSyncTransaction(
-			m.GetID(),
-			m.draftTransaction.Configuration.Sync,
-			m.GetOptions(true)...,
-		)
-
-		// Found any p2p outputs?
-		p2pStatus := SyncStatusSkipped
-		if m.draftTransaction.Configuration.Outputs != nil {
-			for _, output := range m.draftTransaction.Configuration.Outputs {
-				if output.PaymailP4 != nil && output.PaymailP4.ResolutionType == ResolutionTypeP2P {
-					p2pStatus = SyncStatusPending
-				}
-			}
-		}
-		sync.P2PStatus = p2pStatus
-
-		// Use the same metadata
-		sync.Metadata = m.Metadata
-
-		// set this transaction on the sync transaction object. This is needed for the first broadcast
-		sync.transaction = m
-
-		// If all the options are skipped, do not make a new model (ignore the record)
-		if !sync.isSkipped() {
-			m.syncTransaction = sync
-		}
-	}
-
-	// If we are external and the user disabled incoming transaction checking, check outputs
-	if m.isExternal() && !m.Client().IsITCEnabled() {
-		// Check that the transaction has >= 1 known destination
-		if !m.TransactionBase.hasOneKnownDestination(ctx, m.Client(), m.GetOptions(false)...) {
-			return ErrNoMatchingOutputs
-		}
-	}
-
-	// Process the UTXOs
-	if err = m.processUtxos(ctx); err != nil {
-		return err
-	}
-
-	// Set the values from the inputs/outputs and draft tx
-	m.TotalValue, m.Fee = m.getValues()
-
-	// Add values if found
-	if m.TransactionBase.parsedTx != nil {
-		m.NumberOfInputs = uint32(len(m.TransactionBase.parsedTx.Inputs))
-		m.NumberOfOutputs = uint32(len(m.TransactionBase.parsedTx.Outputs))
-	}
-
-	m.DebugLog("end: " + m.Name() + " BeforeCreating hook")
-	m.beforeCreateCalled = true
-	return nil
-}
-
-// AfterCreated will fire after the model is created in the Datastore
-func (m *Transaction) AfterCreated(ctx context.Context) error {
-	m.DebugLog("starting: " + m.Name() + " AfterCreated hook...")
-
-	// Pre-build the options
-	opts := m.GetOptions(false)
-
-	// update the xpub balances
-	for xPubID, balance := range m.XpubOutputValue {
-		// todo: run this in a go routine? (move this into a function on the xpub model?)
-		xPub, err := getXpubWithCache(ctx, m.Client(), "", xPubID, opts...)
-		if err != nil {
-			return err
-		} else if xPub == nil {
-			return ErrMissingRequiredXpub
-		}
-		if err = xPub.incrementBalance(ctx, balance); err != nil {
-			return err
-		}
-	}
-
-	// Update the draft transaction, process broadcasting
-	// todo: go routine (however it's not working, panic in save for missing datastore)
-	if m.draftTransaction != nil {
-		m.draftTransaction.Status = DraftStatusComplete
-		m.draftTransaction.FinalTxID = m.ID
-		if err := m.draftTransaction.Save(ctx); err != nil {
-			return err
-		}
-	}
-
-	// Fire notifications (this is already in a go routine)
-	notify(notifications.EventTypeCreate, m)
-
-	m.DebugLog("end: " + m.Name() + " AfterCreated hook")
-	return nil
-}
-
-// AfterUpdated will fire after the model is updated in the Datastore
-func (m *Transaction) AfterUpdated(_ context.Context) error {
-	m.DebugLog("starting: " + m.Name() + " AfterUpdated hook...")
-
-	// Fire notifications (this is already in a go routine)
-	notify(notifications.EventTypeUpdate, m)
-
-	m.DebugLog("end: " + m.Name() + " AfterUpdated hook")
-	return nil
-}
-
-// AfterDeleted will fire after the model is deleted in the Datastore
-func (m *Transaction) AfterDeleted(_ context.Context) error {
-	m.DebugLog("starting: " + m.Name() + " AfterDelete hook...")
-
-	// Fire notifications (this is already in a go routine)
-	notify(notifications.EventTypeDelete, m)
-
-	m.DebugLog("end: " + m.Name() + " AfterDelete hook")
-	return nil
-}
-
-// ChildModels will get any related sub models
-func (m *Transaction) ChildModels() (childModels []ModelInterface) {
-	// Add the UTXOs if found
-	for index := range m.utxos {
-		childModels = append(childModels, &m.utxos[index])
-	}
-
-	// Add the broadcast transaction record
-	if m.syncTransaction != nil {
-		childModels = append(childModels, m.syncTransaction)
-	}
-
-	return
-}
-
-// processUtxos will process the inputs and outputs for UTXOs
-func (m *Transaction) processUtxos(ctx context.Context) error {
-	// Input should be processed only for outgoing transactions
-	if m.draftTransaction != nil {
-		if err := m.processInputs(ctx); err != nil {
-			return err
-		}
-	}
-
-	return m.processOutputs(ctx)
-}
-
-// processTxOutputs will process the transaction outputs
-func (m *Transaction) processOutputs(ctx context.Context) (err error) {
-	// Pre-build the options
-	opts := m.GetOptions(false)
-	newOpts := append(opts, New())
-	var destination *Destination
-
-	// check all the outputs for a known destination
-	numberOfOutputsProcessed := 0
-	for index := range m.TransactionBase.parsedTx.Outputs {
-		amount := m.TransactionBase.parsedTx.Outputs[index].Satoshis
-
-		// only save outputs with a satoshi value attached to it
-		if amount > 0 {
-
-			txLockingScript := m.TransactionBase.parsedTx.Outputs[index].LockingScript.String()
-			lockingScript := utils.GetDestinationLockingScript(txLockingScript)
-
-			// only Save utxos for known destinations
-			// todo: optimize this SQL SELECT by requesting all the scripts at once (vs in this loop)
-			// todo: how to handle tokens and other non-standard outputs ?
-			if destination, err = m.transactionService.getDestinationByLockingScript(
-				ctx, lockingScript, opts...,
-			); err != nil {
-				return
-			} else if destination != nil {
-
-				// Add value of output to xPub ID
-				if _, ok := m.XpubOutputValue[destination.XpubID]; !ok {
-					m.XpubOutputValue[destination.XpubID] = 0
-				}
-				m.XpubOutputValue[destination.XpubID] += int64(amount)
-
-				utxo, _ := m.client.GetUtxoByTransactionID(ctx, m.ID, uint32(index))
-				if utxo == nil {
-					utxo = newUtxo(
-						destination.XpubID, m.ID, txLockingScript, uint32(index),
-						amount, newOpts...,
-					)
-				}
-				// Append the UTXO model
-				m.utxos = append(m.utxos, *utxo)
-
-				// Add the xPub ID
-				if !utils.StringInSlice(destination.XpubID, m.XpubOutIDs) {
-					m.XpubOutIDs = append(m.XpubOutIDs, destination.XpubID)
-				}
-
-				numberOfOutputsProcessed++
-			}
-		}
 	}
 
 	return
@@ -821,98 +352,6 @@ func (m *Transaction) Display() interface{} {
 	return m
 }
 
-// Migrate model specific migration on startup
-func (m *Transaction) Migrate(client datastore.ClientInterface) error {
-	tableName := client.GetTableName(tableTransactions)
-	if client.Engine() == datastore.MySQL {
-		if err := m.migrateMySQL(client, tableName); err != nil {
-			return err
-		}
-	} else if client.Engine() == datastore.PostgreSQL {
-		if err := m.migratePostgreSQL(client, tableName); err != nil {
-			return err
-		}
-	}
-	
-	err := m.migrateBUMP()
-	if err != nil {
-		return err
-	}
-
-	return client.IndexMetadata(tableName, xPubMetadataField)
-}
-
-// migratePostgreSQL is specific migration SQL for Postgresql
-func (m *Transaction) migratePostgreSQL(client datastore.ClientInterface, tableName string) error {
-	tx := client.Execute(`CREATE INDEX IF NOT EXISTS idx_` + tableName + `_xpub_in_ids ON ` +
-		tableName + ` USING gin (xpub_in_ids jsonb_ops)`)
-	if tx.Error != nil {
-		return tx.Error
-	}
-
-	if tx = client.Execute(`CREATE INDEX IF NOT EXISTS idx_` + tableName + `_xpub_out_ids ON ` +
-		tableName + ` USING gin (xpub_out_ids jsonb_ops)`); tx.Error != nil {
-		return tx.Error
-	}
-
-	return nil
-}
-
-// migrateMySQL is specific migration SQL for MySQL
-func (m *Transaction) migrateMySQL(client datastore.ClientInterface, tableName string) error {
-	idxName := "idx_" + tableName + "_xpub_in_ids"
-	idxExists, err := client.IndexExists(tableName, idxName)
-	if err != nil {
-		return err
-	}
-	if !idxExists {
-		tx := client.Execute("ALTER TABLE `" + tableName + "`" +
-			" ADD INDEX " + idxName + " ( (CAST(xpub_in_ids AS CHAR(64) ARRAY)) )")
-		if tx.Error != nil {
-			m.Client().Logger().Error(context.Background(), "failed creating json index on mysql: "+tx.Error.Error())
-			return nil //nolint:nolintlint,nilerr // error is not needed
-		}
-	}
-
-	idxName = "idx_" + tableName + "_xpub_out_ids"
-	if idxExists, err = client.IndexExists(
-		tableName, idxName,
-	); err != nil {
-		return err
-	}
-	if !idxExists {
-		tx := client.Execute("ALTER TABLE `" + tableName + "`" +
-			" ADD INDEX " + idxName + " ( (CAST(xpub_out_ids AS CHAR(64) ARRAY)) )")
-		if tx.Error != nil {
-			m.Client().Logger().Error(context.Background(), "failed creating json index on mysql: "+tx.Error.Error())
-			return nil //nolint:nolintlint,nilerr // error is not needed
-		}
-	}
-
-	tx := client.Execute("ALTER TABLE `" + tableName + "` MODIFY COLUMN hex longtext")
-	if tx.Error != nil {
-		m.Client().Logger().Error(context.Background(), "failed changing hex type to longtext in MySQL: "+tx.Error.Error())
-		return nil //nolint:nolintlint,nilerr // error is not needed
-	}
-
-	return nil
-}
-
-func (m *Transaction) migrateBUMP() error {
-	ctx := context.Background()
-	txs, err := getTransactionsToCalculateBUMP(ctx, nil, WithClient(m.client))
-	if err != nil {
-		return err
-	}
-	for _, tx := range txs {
-		bump := tx.MerkleProof.ToBUMP()
-		bump.BlockHeight = tx.BlockHeight
-		tx.BUMP = bump
-		tx.Save(ctx)
-	}
-	return nil
-}
-
 // hasOneKnownDestination will check if the transaction has at least one known destination
 //
 // This is used to validate if an external transaction should be recorded into the engine
@@ -961,121 +400,4 @@ func (m *Transaction) RegisterTasks() error {
 		RunEveryPeriod: m.Client().GetTaskPeriod(checkTask),
 		TaskName:       checkTask,
 	})
-}
-
-// processTransactions will process transaction records
-func processTransactions(ctx context.Context, maxTransactions int, opts ...ModelOps) error {
-	queryParams := &datastore.QueryParams{
-		Page:          1,
-		PageSize:      maxTransactions,
-		OrderByField:  "created_at",
-		SortDirection: "asc",
-	}
-
-	conditions := map[string]interface{}{
-		"$or": []map[string]interface{}{{
-			blockHeightField: 0,
-		}, {
-			blockHeightField: nil,
-		}},
-	}
-
-	records := make([]Transaction, 0)
-	err := getModelsByConditions(ctx, ModelTransaction, &records, nil, &conditions, queryParams, opts...)
-	if err != nil {
-		return err
-	} else if len(records) == 0 {
-		return nil
-	}
-
-	txs := make([]*Transaction, 0)
-	for index := range records {
-		records[index].enrich(ModelTransaction, opts...)
-		txs = append(txs, &records[index])
-	}
-
-	for index := range records {
-		if err = processTransaction(
-			ctx, txs[index],
-		); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// processTransaction will process the sync transaction record, or save the failure
-func processTransaction(ctx context.Context, transaction *Transaction) error {
-	txInfo, err := transaction.Client().Chainstate().QueryTransactionFastest(
-		ctx, transaction.ID, chainstate.RequiredOnChain, defaultQueryTxTimeout,
-	)
-	if err != nil {
-		if errors.Is(err, chainstate.ErrTransactionNotFound) {
-			return nil
-		}
-		return err
-	}
-
-	transaction.BlockHash = txInfo.BlockHash
-	transaction.BlockHeight = uint64(txInfo.BlockHeight)
-
-	return transaction.Save(ctx)
-}
-
-// getTransactionsByConditions will get the sync transactions to migrate
-func getTransactionsByConditions(ctx context.Context, conditions map[string]interface{},
-	queryParams *datastore.QueryParams, opts ...ModelOps,
-) ([]*Transaction, error) {
-	if queryParams == nil {
-		queryParams = &datastore.QueryParams{
-			OrderByField:  createdAtField,
-			SortDirection: datastore.SortAsc,
-		}
-	} else if queryParams.OrderByField == "" || queryParams.SortDirection == "" {
-		queryParams.OrderByField = createdAtField
-		queryParams.SortDirection = datastore.SortAsc
-	}
-	
-	// Get the records
-	var models []Transaction
-	if err := getModels(
-		ctx, NewBaseModel(ModelNameEmpty, opts...).Client().Datastore(),
-		&models, conditions, queryParams, defaultDatabaseReadTimeout,
-	); err != nil {
-		if errors.Is(err, datastore.ErrNoResults) {
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	// Loop and enrich
-	txs := make([]*Transaction, 0)
-	for index := range models {
-		models[index].enrich(ModelTransaction, opts...)
-		txs = append(txs, &models[index])
-	}
-
-	return txs, nil
-}
-
-// getTransactionsToMigrateMerklePath will get the transactions where bump should be calculated
-func getTransactionsToCalculateBUMP(ctx context.Context, queryParams *datastore.QueryParams,
-	opts ...ModelOps,
-) ([]*Transaction, error) {
-	// Get the records by status
-	txs, err := getTransactionsByConditions(
-		ctx,
-		map[string]interface{}{
-			bumpField: nil,
-			merkleProofField: map[string]interface{}{
-				"$exists": true,
-			},
-		},
-		queryParams, opts...,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return txs, nil
 }
